refactor(@formatjs/intl-numberformat): let Gazelle own number data helpers
