+++ packages/intl-numberformat/scripts/number-data/BUILD.bazel
@@ -1,0 +1,30 @@
+load("//tools:compile.bzl", "formatjs_library")
+load("//tools:test.bzl", "formatjs_test")
+
+formatjs_library(
+    name = "number-data",
+    srcs = [
+        "number-data.ts",
+        "numbering-system-data.ts",
+        "utils.ts",
+    ],
+    tsconfig_types = ["node"],
+    visibility = ["//packages/intl-numberformat:__subpackages__"],
+    deps = [
+        "//:node_modules/@types/lodash-es",
+        "//:node_modules/@types/node",
+        "//:node_modules/cldr-core",
+        "//:node_modules/fast-xml-parser",
+        "//:node_modules/lodash-es",
+        "//packages/ecma402-abstract/types",
+    ],
+)
+
+formatjs_test(
+    name = "number-data_test",
+    srcs = ["number-data.test.ts"],
+    deps = [
+        ":number-data",
+        "//:node_modules/vitest",
+    ],
+)
