+++ packages/intl-numberformat/scripts/number-data/number-data.ts
