+++ packages/intl-numberformat/scripts/number-data/utils.ts
